# swagger.yaml
openapi: 3.0.0
info:
  description: Sistema de marketing y ventas.
  version: 1.0.0
  title: Sistema de marketing y ventas
servers:
  - url: 'http://localhost:3000'
    description: 'Entorno de desarrollo'
paths:

  /parser/parse_excel:
    post:
      summary: Parsea archivo de Excel
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: archivo procesado correctamente

  /parser/parse_csv:
    post:
      summary: Parsea archivo de CSV
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Archivo procesado correctamente

  /parser/parse_json:
    post:
      summary: Parsear archivo JSON
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Archivo JSON a subir
      responses:
        '200':
          description: Archivo procesado correctamente

  /productos:
    get:
      summary: Obtener productos cargados
      description: Devuelve una lista de todos los productos en el sistema.
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nombre:
                      type: string
                    categoria:
                      type: string
                    proveedor:
                      type: string
                    precio_base:
                      type: integer
        '500':
          description: Error al obtener productos

  /stock:
    get:
      summary: Obtener stock disponible
      description: Devuelve una lista del stock actual con nombre de producto y cantidad.
      responses:
        '200':
          description: Lista de stock
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    producto:
                      type: string
                    cantidad:
                      type: integer
        '500':
          description: Error al obtener stock

  /enviar-reporte:
    post:
      summary: Enviar reporte de stock por email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: ejemplo@correo.com
              required:
                - email
      responses:
        '200':
          description: Reporte enviado correctamente
        '400':
          description: Falta el email en el cuerpo de la petición
        '500':
          description: Error al enviar el reporte